Handle processing config from workflow package
